using-a-native-app-to-build-a-route-optimisation-simulator-hands-on-lab/risk_intelligence_test_instance/requirements.txt
# Risk Intelligence Test Instance Requirements
snowflake-connector-python>=3.0.0
pandas>=1.5.0
numpy>=1.24.0
